+++ library/src/validations/index.ts
@@ -29,3 +29,4 @@ export * from './size/index.ts';
 export * from './startsWith/index.ts';
 export * from './url/index.ts';
 export * from './uuid/index.ts';
+export * from './value/index.ts';